--- linkis-computation-governance/linkis-engineconn/linkis-computation-engineconn/src/test/scala/org/apache/linkis/engineconn/computation/executor/hook/PythonModuleLoadTest.scala
@@ -1,63 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.linkis.engineconn.computation.executor.hook
-
-import org.apache.linkis.udf.entity.PythonModuleInfoVO
-import org.junit.jupiter.api.Test
-
-import java.util
-
-class PythonModuleLoadTest {
-  /**
-   * 测试getEngineType方法，确保返回正确的引擎类型。
-   */
-  @Test def testGetEngineType(): Unit = {
-    val pythonModuleLoad: PythonModuleLoad = new PythonSparkEngineHook() {
-      override protected def getEngineType = "Spark"
-    }
-  }
-
-  /**
-   * 测试constructCode方法，确保构建的代码字符串正确。
-   */
-  @Test def testConstructCode(): Unit = {
-    val pythonModuleLoad: PythonModuleLoad = new PythonSparkEngineHook() {
-      protected def constructCode(pythonModuleInfo: Nothing): String = "import "
-    }
-    val moduleInfo = new Nothing("numpy", "/path/to/numpy")
-    val expectedCode = "import numpy"
-  }
-
-  /**
-   * 测试loadPythonModules方法，确保模块加载逻辑正确。
-   */
-  @Test def testLoadPythonModules(): Unit = {
-    val pythonModuleLoad: PythonModuleLoad = new PythonSparkEngineHook() {
-      override protected def getEngineType = "Spark"
-
-      protected def constructCode(pythonModuleInfo: Nothing): String = "import "
-    }
-    val moduleInfoList = new util.ArrayList[PythonModuleInfoVO]()
-    moduleInfoList.add(new Nothing("numpy", "/path/to/numpy"))
-    moduleInfoList.add(new Nothing("pandas", "/path/to/pandas"))
-    // val labels = new Array[Label[_]]
-    // pythonModuleLoad.loadPythonModules(labels)
-    // 如果loadPythonModules方法有副作用，例如修改外部状态或调用其他方法，
-    // 那么这里应该添加相应的断言或验证。
-  }
-}
