+++ linkis-computation-governance/linkis-manager/linkis-manager-common/src/main/java/org/apache/linkis/manager/common/entity/node/EngineNode.java
@@ -34,4 +34,8 @@ public interface EngineNode extends AMNode, RMNode, LabelNode {
   String getEcMetrics();
 
   void setEcMetrics(String metrics);
+
+  String getParams();
+
+  void setParams(String params);
 }
