+++ linkis-computation-governance/linkis-manager/linkis-manager-common/src/main/java/org/apache/linkis/manager/common/entity/persistence/PersistenceNode.java
@@ -35,6 +35,7 @@ public class PersistenceNode {
   private Date createTime;
   private String updator;
   private String creator;
+  private String params;
 
   public String getMark() {
     return mark;
@@ -122,5 +123,17 @@ public class PersistenceNode {
 
   public void setCreator(String creator) {
     this.creator = creator;
+  }
+
+  public void setId(int id) {
+    this.id = id;
+  }
+
+  public String getParams() {
+    return params;
+  }
+
+  public void setParams(String params) {
+    this.params = params;
   }
 }
