+++ linkis-dist/package/db/upgrade/1.4.1_schema/mysql/linkis_ddl.sql
@@ -73,3 +73,5 @@ ALTER TABLE linkis_cg_tenant_label_config ADD COLUMN is_valid varchar(1) CHARSET
 ALTER TABLE linkis_cg_manager_service_instance_metrics ADD COLUMN description varchar(256) CHARSET utf8mb4 COLLATE utf8mb4_bin DEFAULT '';
 ALTER TABLE linkis_ps_bml_resources_task ADD CONSTRAINT  uniq_rid_version UNIQUE (`resource_id`, `version`);
 ALTER TABLE linkis_cg_ec_resource_info_record ADD UNIQUE INDEX uniq_sinstance_status_cuser_ctime (`service_instance`, `status`, `create_user`, `create_time`);
+
+ALTER TABLE linkis_cg_manager_service_instance ADD COLUMN  params text COLLATE utf8_bin DEFAULT NULL;
